directories into packages
