+++ .gitignore
@@ -10,3 +10,6 @@
 
 # Built Visual Studio Code Extensions
 *.vsix
+
+# Python compiled files
+__pycache__/
