+++ grammar/.antlr/DepotParser.java
@@ -1,4 +1,4 @@
-// Generated from /Users/jakubwadrzyk/Documents/Programing/Python/Kompilatory/Depot/grammar/DepotParser.g4 by ANTLR 4.9.2
+// Generated from /Users/jakubwadrzyk/Documents/Programing/Python/Kompilatory/DepotGrammar/grammar/DepotParser.g4 by ANTLR 4.9.2
 import org.antlr.v4.runtime.atn.*;
 import org.antlr.v4.runtime.dfa.DFA;
 import org.antlr.v4.runtime.*;
