+++ grammar/__init__.py
@@ -1,0 +1,3 @@
+from .DepotLexer import DepotLexer
+from .DepotParser import DepotParser
+from .DepotParserListener import DepotParserListener
